The Beginning is the End and the End is the Beginning

+++ src/main/java/ConsomiTounsi/Controller/ClaimAndAd/ClaimController.java
@@ -63,8 +63,11 @@ public class ClaimController {
 	}
 	
 	
+
+	
+	
 	@GetMapping("/Bp/{Bp}")
-	public int claimsNumber(@PathVariable("Bp") int Bp){
+	public int claimsNumber(@PathVariable("Bp") Long Bp){
 		return claimManagerI.claimsNumber(Bp);
 	}
